Changed Coordinate to use floats and fixed indents

+++ core/src/com/mikroysoft/game/Alien.java
@@ -24,8 +24,8 @@ public class Alien implements IRenderable {
         return this.position;
     }
     public void Run(){
-        position.x += Integer.signum(position.y + (int)(Math.random()))*2;
-        position.y += Integer.signum(position.y + (int)(Math.random()))*2;
+        position.x += Integer.signum((int)position.y + (int)(Math.random()))*2;
+        position.y += Integer.signum((int)position.y + (int)(Math.random()))*2;
         direction = (float) Math.toDegrees(Math.atan2((position.y +(Math.random()* 10 + 1)) * -1,  position.x - (Math.random()* 10 + 1))) +45;
         System.out.println(position.x);
     }
